Update README.md
Breaking Bad is a series about a character that is overqualified in his job as a chemistry teacher. One day as he discover he has cancer, he wanted to leave behind money to support his family. Thus, he began breaking bad.

Hell's Kitchen is a series of episode regarding the contestants competing with Gordon Ramsey as the Chef that teaches and oversees the contestant.

Cowboy Bebop is an anime series about a bounty hunter going from place to place. 

Re:zero is an anime about character transported in an fantasy world. The main character discovers his ability to travel back in time if he died. His journey towards helping the people he cares about while learning more about life is the main theme of the series